requirements.txt
streamlit==1.29.0
supabase==2.0.3
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.0.1
